[NO_TICKET] Reorganize exports (#953)
* Add index.js files in each component folder, update entry index.js

* Separate out mask helper functions

* Separate out getMonthNames

* Separate out mask helper function tests

* Export maskValue

* Update tests

* fix sort-import issue in FormControl
