+++ packages/design-system/src/components/Alert/index.js
@@ -1,0 +1,1 @@
+export { default as Alert } from './Alert';
