+++ packages/design-system/src/components/Autocomplete/index.js
@@ -1,0 +1,1 @@
+export { default as Autocomplete } from './Autocomplete';
