+++ packages/design-system/src/components/Button/index.js
@@ -1,0 +1,1 @@
+export { default as Button } from './Button';
