+++ packages/design-system/src/components/ChoiceList/index.js
@@ -1,0 +1,2 @@
+export { default as ChoiceList } from './ChoiceList';
+export { default as Choice } from './Choice';
