+++ packages/design-system/src/components/DateField/index.js
@@ -1,0 +1,1 @@
+export { default as DateField } from './DateField';
