+++ packages/design-system/src/components/Dialog/index.js
@@ -1,0 +1,1 @@
+export { default as Dialog } from './Dialog';
