+++ packages/design-system/src/components/Dropdown/index.js
@@ -1,0 +1,1 @@
+export { default as Dropdown } from './Dropdown';
