+++ packages/design-system/src/components/FormLabel/index.js
@@ -1,0 +1,1 @@
+export { default as FormLabel } from './FormLabel';
