+++ packages/design-system/src/components/MonthPicker/index.js
@@ -1,0 +1,2 @@
+export { default as MonthPicker } from './MonthPicker';
+export { getMonthNames } from './getMonthNames';
