+++ packages/design-system/src/components/Review/index.js
@@ -1,0 +1,1 @@
+export { default as Review } from './Review';
