+++ packages/design-system/src/components/SkipNav/index.js
@@ -1,0 +1,1 @@
+export { default as SkipNav } from './SkipNav';
