+++ packages/design-system/src/components/Spinner/index.js
@@ -1,0 +1,1 @@
+export { default as Spinner } from './Spinner';
