+++ packages/design-system/src/components/StepList/index.js
@@ -1,0 +1,1 @@
+export { default as StepList } from './StepList';
