+++ packages/design-system/src/components/Table/index.js
@@ -1,0 +1,6 @@
+export { default as Table } from './Table';
+export { default as TableBody } from './TableBody';
+export { default as TableCaption } from './TableCaption';
+export { default as TableCell } from './TableCell';
+export { default as TableHead } from './TableHead';
+export { default as TableRow } from './TableRow';
