+++ packages/design-system/src/components/Tabs/index.js
@@ -1,0 +1,2 @@
+export { default as Tabs } from './Tabs';
+export { default as TabPanel } from './TabPanel';
