+++ packages/design-system/src/components/TextField/index.js
@@ -1,0 +1,2 @@
+export { default as TextField } from './TextField';
+export { maskValue, unmaskValue } from './maskHelpers';
