+++ packages/design-system/src/components/Tooltip/index.js
@@ -1,0 +1,2 @@
+export { default as Tooltip } from './Tooltip';
+export { default as TooltipIcon } from './TooltipIcon';
