+++ packages/design-system/src/components/UsaBanner/index.js
@@ -1,0 +1,1 @@
+export { default as UsaBanner } from './UsaBanner';
