+++ packages/design-system/src/components/VerticalNav/index.js
@@ -1,0 +1,1 @@
+export { default as VerticalNav } from './VerticalNav';
