+++ packages/design-system/src/components/index.js
@@ -2,34 +2,26 @@
  * index.js - JS entry point
  */
 
-export { default as Alert } from './Alert/Alert';
-export { default as Autocomplete } from './Autocomplete/Autocomplete';
-export { default as Badge } from './Badge/Badge';
-export { default as Button } from './Button/Button';
-export { default as Choice } from './ChoiceList/Choice';
-export { default as ChoiceList } from './ChoiceList/ChoiceList';
-export { default as DateField } from './DateField/DateField';
-export { default as Dialog } from './Dialog/Dialog';
-export { default as Dropdown } from './Dropdown/Dropdown';
-export { default as FormLabel } from './FormLabel/FormLabel';
-export { default as HelpDrawer } from './HelpDrawer/HelpDrawer';
-export { default as HelpDrawerToggle } from './HelpDrawer/HelpDrawerToggle';
-export { default as MonthPicker, getMonthNames } from './MonthPicker/MonthPicker';
-export { default as Review } from './Review/Review';
-export { default as SkipNav } from './SkipNav/SkipNav';
-export { default as Spinner } from './Spinner/Spinner';
-export { default as StepList } from './StepList/StepList';
-export { default as Table } from './Table/Table';
-export { default as TableBody } from './Table/TableBody';
-export { default as TableCaption } from './Table/TableCaption';
-export { default as TableCell } from './Table/TableCell';
-export { default as TableHead } from './Table/TableHead';
-export { default as TableRow } from './Table/TableRow';
-export { default as Tabs } from './Tabs/Tabs';
-export { default as TabPanel } from './Tabs/TabPanel';
-export { default as TextField, unmaskValue } from './TextField/TextField';
-export { default as Tooltip } from './Tooltip/Tooltip';
-export { default as TooltipIcon } from './Tooltip/TooltipIcon';
-export { default as UsaBanner } from './UsaBanner/UsaBanner';
-export { default as VerticalNav } from './VerticalNav/VerticalNav';
+export * from './Alert';
+export * from './Autocomplete';
+export * from './Badge';
+export * from './Button';
+export * from './ChoiceList';
+export * from './DateField';
+export * from './Dialog';
+export * from './Dropdown';
+export * from './FormLabel';
+export * from './HelpDrawer';
+export * from './MonthPicker';
+export * from './Review';
+export * from './SkipNav';
+export * from './Spinner';
+export * from './StepList';
+export * from './Table';
+export * from './Tabs';
+export * from './TextField';
+export * from './Tooltip';
+export * from './UsaBanner';
+export * from './VerticalNav';
+
 export * from './flags';
